+++ tools/batools/assetsmakefile.py
@@ -134,7 +134,7 @@ def _get_py_targets(
             elif proot.startswith(f'{ASSETS_SRC}/windows/Win32'):
                 in_subset = 'private-windows-Win32'
             elif proot.startswith(
-                f'src/external/python-apple/macos/Python.xcframework/'
+                f'src/external/python-apple-old/macos/Python.xcframework/'
                 f'macos-arm64_x86_64/Python.framework/'
                 f'Versions/{PYVER}/lib/python{PYVER}'
             ):
@@ -245,20 +245,20 @@ def _get_py_targets_subset(
     # build/assets/pylib-apple-mac
     elif subset == 'private-apple-mac':
         src = (
-            f'src/external/python-apple/macos/Python.xcframework/'
+            f'src/external/python-apple-old/macos/Python.xcframework/'
             f'macos-arm64_x86_64/Python.framework/'
             f'Versions/{PYVER}/lib/python{PYVER}'
         )
-        dst = f'{BUILD_DIR}/python-apple/macos/pylib'
+        dst = f'{BUILD_DIR}/python-apple-old/macos/pylib'
         copyrule = (
-            f'$(BUILD_DIR)/python-apple/macos/pylib/%.py :'
-            f' $(SRC_DIR)/external/python-apple/macos/Python.xcframework/'
+            f'$(BUILD_DIR)/python-apple-old/macos/pylib/%.py :'
+            f' $(SRC_DIR)/external/python-apple-old/macos/Python.xcframework/'
             f'macos-arm64_x86_64/Python.framework/'
             f'Versions/{PYVER}/lib/python{PYVER}/%.py'
         )
         copyrule_so = (
-            f'$(BUILD_DIR)/python-apple/macos/pylib/%.so :'
-            f' $(SRC_DIR)/external/python-apple/macos/Python.xcframework/'
+            f'$(BUILD_DIR)/python-apple-old/macos/pylib/%.so :'
+            f' $(SRC_DIR)/external/python-apple-old/macos/Python.xcframework/'
             f'macos-arm64_x86_64/Python.framework/'
             f'Versions/{PYVER}/lib/python{PYVER}/%.so'
         )